Use a custom logger
Instead of directly writing with `log.Printf`, this allows the
user to specify a custom logger via an `io.Writer`.  If none is
specified, then no logging will be done.

Historically, this would log an "API rate limit hit" message every
time it needed to sleep a bit due to rate limiting, which, in a
production appliction, is frequent and cluttered the logs.