Sleep 1 second longer than required
When sleeping in order to wait for the rate limit to pass, this
adds one additional second to the sleep time so that we account
for any integer rounding issues.  We don't want to try again before
the server thinks it's time to because it'll hit us with a 5-minute
sleep.