Update hibp.go
Switch version number back to v1.x.x since I really don't want to mess with Go's v2 mechanic (some Go tools are not v2 aware and therefore we would need to branch the new version to a v2 branch or subdir, which I really don't wanna do). This breaks with semver semantics, but it is, what it is.